Merged from gioblu master

+++ examples/Local/OverSamplingHalfDuplexNoAcknowledge/Device1/Device1.ino
@@ -18,7 +18,7 @@ void setup() {
   Serial.begin(115200);
 };
 
-void receiver_function(uint8_t *payload, uint8_t length, const PacketInfo &packet_info) {
+void receiver_function(uint8_t *payload, uint16_t length, const PacketInfo &packet_info) {
  if((char)payload[0] == 'B') {
     digitalWrite(13, HIGH);
     delay(5);
