+++ examples/Local/SendArbitraryValues/Receiver/Receiver.ino
@@ -19,7 +19,7 @@ void setup() {
   Serial.begin(115200);
 };
 
-void receiver_function(uint8_t *payload, uint8_t length, const PacketInfo &packet_info) {
+void receiver_function(uint8_t *payload, uint16_t length, const PacketInfo &packet_info) {
   Serial.print("RX:");
   Serial.print(" Device id: ");
   Serial.print(packet_info.receiver_id);
