+++ examples/Network/SendAndReceive/Device2/Device2.ino
@@ -20,7 +20,7 @@ void setup() {
   Serial.begin(115200);
 };
 
-void receiver_function(uint8_t *payload, uint8_t length, const PacketInfo &packet_info) {
+void receiver_function(uint8_t *payload, uint16_t length, const PacketInfo &packet_info) {
  if((char)payload[0] == 'B') {
     if(!bus.packets[packet].state)
       packet = bus.reply("B", 1);
